requirements.txt
anthropic>=0.18.0
pyyaml>=6.0
click>=8.0.0
rich>=13.0.0
gitpython>=3.1.0
jinja2>=3.0.0
python-slugify>=8.0.0
tqdm>=4.65.0
